Changed type of compute invocation and group counting to uint32

+++ VitraeEngine/include/Vitrae/Renderers/OpenGL/ShaderCompilation.hpp
@@ -24,9 +24,9 @@ class CompiledGLSLShader : public dynasma::PolymorphicBase
   public:
     struct ComputeCompilationSpec
     {
-        PropertyGetter<std::size_t> invocationCountX;
-        PropertyGetter<std::size_t> invocationCountY;
-        PropertyGetter<std::size_t> invocationCountZ;
+        PropertyGetter<std::uint32_t> invocationCountX;
+        PropertyGetter<std::uint32_t> invocationCountY;
+        PropertyGetter<std::uint32_t> invocationCountZ;
         glm::uvec3 groupSize;
         bool allowOutOfBoundsCompute;
     };
@@ -86,9 +86,9 @@ class CompiledGLSLShader : public dynasma::PolymorphicBase
         dynasma::FirmPtr<Method<ShaderTask>> mp_computeMethod;
         dynasma::FirmPtr<const PropertyList> mp_desiredResults;
         ComponentRoot *mp_root;
-        PropertyGetter<std::size_t> m_invocationCountX;
-        PropertyGetter<std::size_t> m_invocationCountY;
-        PropertyGetter<std::size_t> m_invocationCountZ;
+        PropertyGetter<std::uint32_t> m_invocationCountX;
+        PropertyGetter<std::uint32_t> m_invocationCountY;
+        PropertyGetter<std::uint32_t> m_invocationCountZ;
         glm::uvec3 m_groupSize;
         bool m_allowOutOfBoundsCompute;
         std::size_t m_hash;
@@ -97,17 +97,17 @@ class CompiledGLSLShader : public dynasma::PolymorphicBase
         ComputeShaderParams(ComponentRoot &root,
                             dynasma::FirmPtr<Method<ShaderTask>> p_computeMethod,
                             dynasma::FirmPtr<const PropertyList> p_desiredResults,
-                            PropertyGetter<std::size_t> invocationCountX,
-                            PropertyGetter<std::size_t> invocationCountY,
-                            PropertyGetter<std::size_t> invocationCountZ, glm::uvec3 groupSize,
+                            PropertyGetter<std::uint32_t> invocationCountX,
+                            PropertyGetter<std::uint32_t> invocationCountY,
+                            PropertyGetter<std::uint32_t> invocationCountZ, glm::uvec3 groupSize,
                             bool allowOutOfBoundsCompute);
 
         inline auto getComputeMethodPtr() const { return mp_computeMethod; }
         inline auto getDesiredResultsPtr() const { return mp_desiredResults; }
         inline ComponentRoot &getRoot() const { return *mp_root; }
-        inline PropertyGetter<std::size_t> getInvocationCountX() const { return m_invocationCountX; }
-        inline PropertyGetter<std::size_t> getInvocationCountY() const { return m_invocationCountY; }
-        inline PropertyGetter<std::size_t> getInvocationCountZ() const { return m_invocationCountZ; }
+        inline PropertyGetter<std::uint32_t> getInvocationCountX() const { return m_invocationCountX; }
+        inline PropertyGetter<std::uint32_t> getInvocationCountY() const { return m_invocationCountY; }
+        inline PropertyGetter<std::uint32_t> getInvocationCountZ() const { return m_invocationCountZ; }
         inline glm::uvec3 getGroupSize() const { return m_groupSize; }
         inline bool getAllowOutOfBoundsCompute() const { return m_allowOutOfBoundsCompute; }
